requirements.txt
slack-sdk
praw
requests